requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.0
idna==3.10
PyJWT==1.7.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
